requirements.txt
Django==2.2.12
django-extensions==2.2.9
djangorestframework==3.10.3
ipdb==0.13.2
ipython==7.14.0
